pdf-diff-backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-Levenshtein
